requirements.txt
streamlit
azure-ai-textanalytics
python-dotenv
requests
azure-cognitiveservices-vision-computervision
azure-cognitiveservices-speech
openai
PyMuPDF
audio-recorder-streamlit